Split the tool-packaging loader into focused modules
loader.ts had grown to ~1930 lines spanning four concerns. Decompose it,
leaving loader.ts as the author-code loader (createToolLoader + the
interchange-entry import/eval helpers) at ~940 lines:

- loader-internal.ts -- the shared failure type (ToolLoaderError), the
  fetch byte/time caps, and the small cross-cutting helpers. Split out
  so the two concern modules and loader.ts can all depend on them
  without an import cycle.
- registry-fetch.ts -- building the npm-registry-fetch options, deriving
  a default tarball URL, and reading a response body under a byte cap.
- store-layout.ts -- materializeClosure plus the eval-free closure
  materialization and per-instance store layout it drives (cache
  extract, range resolution, hardlink layout, direct-dependency
  symlinks).

loader.ts re-exports every moved public member (ToolLoaderError, the
DEFAULT_* caps, buildRegistryFetchOpts, readResponseWithLimit,
materializeClosure, storeEntryDir) so `./loader` and package-root
consumers are unaffected. The only cross-module cycle is a type-only
import of the materializeClosure arg/result shapes back from loader.ts,
which has no runtime effect.
